magic_enum: Customize enums where they're declared
This necessitates dependencies on `magic_enum` in more
places but avoids obscure hard-to-debug errors, such
as the one seen in the DOS port PR.
