+++ Source/monstdat.cpp
@@ -29,18 +29,6 @@
 #include "textdat.h"
 #include "utils/language.h"
 
-template <>
-struct magic_enum::customize::enum_range<devilution::_monster_id> {
-	static constexpr int min = devilution::MT_INVALID;
-	static constexpr int max = devilution::NUM_DEFAULT_MTYPES;
-};
-
-template <>
-struct magic_enum::customize::enum_range<devilution::monster_resistance> {
-	static constexpr int min = 0;
-	static constexpr int max = 128;
-};
-
 namespace devilution {
 
 namespace {
