+++ Source/textdat.h
@@ -5,11 +5,11 @@
  */
 #pragma once
 
-#include <cstddef>
 #include <cstdint>
 #include <string>
 
 #include <expected.hpp>
+#include <magic_enum/magic_enum.hpp>
 
 #include "sound_effect_enums.h"
 
@@ -440,3 +440,9 @@ tl::expected<_speech_id, std::string> ParseSpeechId(std::string_view value);
 void LoadTextData();
 
 } // namespace devilution
+
+template <>
+struct magic_enum::customize::enum_range<devilution::_speech_id> {
+	static constexpr int min = devilution::TEXT_NONE;
+	static constexpr int max = devilution::NUM_DEFAULT_TEXT_IDS;
+};
